Adding extra class option to tagsinput wrapper div
If you have multiple and an unknown number of tagsinput instances on a page you cannot use IDs to style them. This commit adds extraClass as an option for a user to specify a class name.

In our case, this has come in useful when we have the tagsinput on two different element type on our page. We initialise it once with one extraClass and again with a different extraClass, enabling us to have control of positional style differences between the two input fields.